Split config.js into multiple files
