+++ src/assets/README.md
@@ -39,7 +39,7 @@
 # 张人大 · 轻量级网站
 
 - **作者**: 张人大
-- **最后更新**: August 05, 2025, 21:15 (UTC+08:00)
+- **最后更新**: August 06, 2025, 01:25 (UTC+08:00)
 
 ---
 
@@ -72,7 +72,7 @@
 ```bash
 src/
 ├── assets/
-├── constants/           # 路径常量与 API 端点
+├── constants/           # 站点元信息、资源路径与应用设置
 ├── features/            # 按业务划分的模块
 │   ├── chat/
 │   │   ├── components/
