+++ src/components/forms/ContactForm/ContactForm.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { CONTACT_FORM_ENDPOINT } from '../../../config.js';
+import { CONTACT_FORM_ENDPOINT } from '../../../constants/index.js';
 import { useLanguage } from '../../providers';
 import { LocalizedSection } from '../../ui';
