+++ src/components/forms/LoginForm/LoginForm.jsx
@@ -1,5 +1,10 @@
 import { useState, useEffect } from 'react';
-import { HOME_PAGE_PATH, REGISTER_PAGE_PATH, LOADING_TEXT, AUTH_TIMINGS } from '../../../config.js';
+import {
+  HOME_PAGE_PATH,
+  REGISTER_PAGE_PATH,
+  LOADING_TEXT,
+  AUTH_TIMINGS
+} from '../../../constants/index.js';
 import { useLanguage } from '../../providers';
 import { LOGIN_CONTENT } from '../../../content';
 import { LocalizedSection } from '../../ui';
