+++ src/components/sections/AboutContent.jsx
@@ -4,7 +4,7 @@ import {
   IMAGE_PATHS,
   RESUME_EN_DOWNLOAD,
   RESUME_ZH_DOWNLOAD
-} from '../../config.js';
+} from '../../constants/index.js';
 import { MAIN_HERO } from '../../data';
 import { ResponsiveHero } from '../ui';
 import ContactSection from './ContactSection.jsx';
