+++ src/components/ui/data-display/SocialIcons.jsx
@@ -1,5 +1,10 @@
 import { useState, useRef, useEffect } from 'react';
-import { SITE_BASE_URL, SOCIAL_ICON_PATHS, SOCIAL_LINKS, IMAGE_PATHS } from '../../../config.js';
+import {
+  SITE_BASE_URL,
+  SOCIAL_ICON_PATHS,
+  SOCIAL_LINKS,
+  IMAGE_PATHS
+} from '../../../constants/index.js';
 
 function SocialIcon({ href, id, src, alt, ariaLabel }) {
   const [loaded, setLoaded] = useState(false);
