+++ src/hooks/useMarkdownPipeline.js
@@ -1,5 +1,5 @@
 import { useEffect, useRef } from 'react';
-import { UI_DURATIONS } from '../config.js';
+import { UI_DURATIONS } from '../constants/index.js';
 
 export function showHint(msg, duration = UI_DURATIONS.HINT) {
   const hint = document.createElement('div');
