+++ src/layouts/BaseLayout.astro
@@ -16,7 +16,7 @@ import {
   TWITTER_HANDLE,
   LANG_STORAGE_KEY,
   THEME_STORAGE_KEY
-} from '../config.js';
+} from '../constants/index.js';
 import '../styles/theme.css';
 
 const {
