+++ src/pages/certifications.astro
@@ -8,7 +8,7 @@ import {
   CERTIFICATIONS_DESCRIPTION,
   CERTIFICATIONS_IMAGE_ALT,
   OG_IMAGE_DIMENSIONS
-} from '../config.js';
+} from '../constants/index.js';
 import { CERTIFICATIONS_CONTENT } from '../content';
 import { CertificationsEffects } from '../components/sections';
 import { getCurrentLang } from '../utils';
