+++ src/pages/deepseek_chat.astro
@@ -3,7 +3,12 @@ import { BaseLayout } from '../layouts';
 import { DeepseekChatContent } from '../components/chat';
 import { DEEPSEEK_CHAT_CONTENT } from '../content';
 import { getCurrentLang } from '../utils';
-import { CHAT_PAGE_PATH, SITE_BASE_URL, STYLE_PATHS, VIEWPORT_ZOOM_ALLOWED } from '../config.js';
+import {
+  CHAT_PAGE_PATH,
+  SITE_BASE_URL,
+  STYLE_PATHS,
+  VIEWPORT_ZOOM_ALLOWED
+} from '../constants/index.js';
 import '../styles/components/deepseek_chat.css';
 
 const langDefault = getCurrentLang();
