+++ src/pages/docs.astro
@@ -9,7 +9,7 @@ import {
   PAGE_TITLES,
   DOCS_DESCRIPTION,
   LOADING_TEXT
-} from '../config.js';
+} from '../constants/index.js';
 import { DocsEffects } from '../components/sections';
 import { getCurrentLang } from '../utils';
