+++ src/pages/index.astro
@@ -7,7 +7,7 @@ import {
   SITE_DESCRIPTION_EN,
   HOME_PAGE_PATH,
   STYLE_PATHS
-} from '../config.js';
+} from '../constants/index.js';
 import { getCurrentLang } from '../utils';
 
 const langDefault = getCurrentLang();
