+++ src/pages/login.astro
@@ -7,7 +7,7 @@ import {
   LOGIN_PAGE_PATH,
   PAGE_TITLES,
   LOGIN_DESCRIPTION
-} from '../config.js';
+} from '../constants/index.js';
 import '../styles/components/login.css';
 import { getCurrentLang } from '../utils';
