+++ src/pages/register.astro
@@ -7,7 +7,7 @@ import {
   REGISTER_PAGE_PATH,
   PAGE_TITLES,
   REGISTER_DESCRIPTION
-} from '../config.js';
+} from '../constants/index.js';
 import '../styles/components/register.css';
 import { getCurrentLang } from '../utils';
